second commit

+++ src/controllers/user.controller.ts
@@ -1,23 +1,35 @@
 import { UserDocument } from "./../models/user.model";
 import { Request, Response } from "express";
-import UserModel from "../models/user.model";
 import * as UserService from "../services/user.service";
+import { isValidObjectId } from "mongoose";
 
 export const GetAll = async (req: Request, res: Response) => {
   try {
-    const result = await UserService.GetAll();
-    return res.status(200).send({ data: result });
-  } catch (error) {
+    const results = await UserService.GetAll();
+    const data = results.map((result) => {
+      return { username: result.username };
+    });
+
+    return res.status(200).send({ data: data });
+  } catch (error: unknown) {
     console.log(error);
     return res.status(400).send({ error });
   }
 };
 
-export const GetById = async (req: Request, res: Response) => {
+export const GetById = async (req: Request<{ id: string }>, res: Response) => {
   try {
+    if (!isValidObjectId(req.params.id)) {
+      return res.status(400).send({ message: "User Id is not valid" });
+    }
+
     const result = await UserService.GetById(req.params.id);
-    return res.status(200).send({ data: result });
-  } catch (error) {
+    if (result == null) {
+      return res.status(404).send({ message: "User did not exists" });
+    }
+
+    return res.status(200).send({ data: result.username });
+  } catch (error: unknown) {
     console.log(error);
     return res.status(400).send({ error });
   }
@@ -37,20 +49,47 @@ export const Create = async (
     await user.save();
 
     return res.status(201).send({ data: user });
-  } catch (error) {
+  } catch (error: unknown) {
     console.log(error);
     return res.status(400).send({ error });
   }
 };
 
-export const Delete = async (req: Request, res: Response) => {
+export const Delete = async (req: Request<{ id: string }>, res: Response) => {
   try {
+    if (!isValidObjectId(req.params.id)) {
+      return res.status(400).send({ message: "User Id is not valid" });
+    }
+
     const user = await UserService.Delete(req.params.id);
     if (user == null) {
       return res.status(404).send({ message: "User did not exists" });
     }
     return res.status(200).send({ user });
-  } catch (error) {
+  } catch (error: unknown) {
+    console.log(error);
+    return res.status(400).send({ error });
+  }
+};
+
+export const Update = async (
+  req: Request<{ id: string }, {}, Pick<UserDocument, "password">>,
+  res: Response
+) => {
+  try {
+    if (!isValidObjectId(req.params.id)) {
+      return res.status(400).send({ message: "User Id is not valid" });
+    }
+
+    const user = await UserService.Update(req.params.id, req.body.password);
+    if (user == null) {
+      return res.status(404).send({ message: "User did not exists" });
+    }
+
+    return res
+      .status(200)
+      .send({ data: { username: user.username, password: user.password } });
+  } catch (error: unknown) {
     console.log(error);
     return res.status(400).send({ error });
   }
